Refactor: restructure projects
