+++ MangaRipper.Plugin.Batoto/packages.config
