+++ MangaRipper.Plugin.MangaFox/packages.config
