+++ MangaRipper.Plugin.MangaHere/packages.config
