+++ MangaRipper.Plugin.MangaReader/MangaReaderService.cs
@@ -9,13 +9,12 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace MangaRipper.MangaHere
+namespace MangaRipper.Plugin.MangaReader
 {
-
     /// <summary>
-    /// Support find chapters and images from MangaHere
+    /// Support find chapters and images from MangaReader
     /// </summary>
-    public class MangaHereService : IMangaService
+    public class MangaReaderService : IMangaService
     {
         private static Logger logger = LogManager.GetCurrentClassLogger();
 
@@ -26,7 +25,14 @@ namespace MangaRipper.MangaHere
             progress.Report(0);
             // find all chapters in a manga
             string input = await downloader.DownloadStringAsync(manga);
-            var chaps = parser.ParseGroup("<a class=\"color_0077\" href=\"(?<Value>http://[^\"]+)\"[^<]+>(?<Name>[^<]+)</a>", input, "Name", "Value");
+            var chaps = parser.ParseGroup("<a href=\"(?<Value>[^\"]+)\">(?<Name>[^<]+)</a> :", input, "Name", "Value");
+            // reverse chapters order and remove duplicated chapters in latest section
+            chaps = chaps.Reverse().GroupBy(x => x.Url).Select(g => g.First()).ToList();
+            // transform pages link
+            chaps = chaps.Select(c =>
+            {
+                return new Chapter(c.Name, new Uri(new Uri(manga), c.Url).AbsoluteUri);
+            }).ToList();
             progress.Report(100);
             return chaps;
         }
@@ -38,7 +44,7 @@ namespace MangaRipper.MangaHere
 
             // find all pages in a chapter
             string input = await downloader.DownloadStringAsync(chapter.Url);
-            var pages = parser.Parse(@"<option value=""(?<Value>[^""]+)"" (|selected=""selected"")>\d+</option>", input, "Value");
+            var pages = parser.Parse(@"<option value=""(?<Value>[^""]+)""(| selected=""selected"")>\d+</option>", input, "Value");
 
             // transform pages link
             pages = pages.Select(p =>
@@ -54,20 +60,20 @@ namespace MangaRipper.MangaHere
                 int i = Convert.ToInt32(f * 100);
                 progress.Report(i);
             }), cancellationToken);
-            var images = parser.Parse("<img src=\"(?<Value>[^\"]+)\" onload=", pageData, "Value");
+            var images = parser.Parse(@"<img id=""img"" width=""\d+"" height=""\d+"" src=""(?<Value>[^""]+)""", pageData, "Value");
 
             return images;
         }
 
         public SiteInformation GetInformation()
         {
-            return new SiteInformation("MangaHere", "http://www.mangahere.co", "English");
+            return new SiteInformation("MangaReader", "http://www.mangareader.net", "English");
         }
 
         public bool Of(string link)
         {
             var uri = new Uri(link);
-            return uri.Host.Equals("www.mangahere.co");
+            return uri.Host.Equals("www.mangareader.net");
         }
     }
 }
