+++ MangaRipper.Plugin.MangaReader/packages.config
