+++ MangaRipper.Plugin.MangaShare/MangaRipper.Plugin.MangaShare.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F5D529B4-CBB6-46E2-9D32-25CEF8365318}</ProjectGuid>
+    <ProjectGuid>{A156FB0B-1E4A-4DB0-B944-D23413CA8D58}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MangaRipper.MangaFox</RootNamespace>
-    <AssemblyName>MangaRipper.MangaFox</AssemblyName>
+    <RootNamespace>MangaRipper.Plugin.MangaShare</RootNamespace>
+    <AssemblyName>MangaRipper.Plugin.MangaShare</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -16,7 +16,7 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>..\MangaRipper\bin\Debug\plugins\</OutputPath>
+    <OutputPath>..\MangaRipper\bin\Debug\Plugins\</OutputPath>
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -44,7 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MangaFoxService.cs" />
+    <Compile Include="MangaShareService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
