+++ MangaRipper.Plugin.MangaShare/packages.config
