+++ MangaRipper.Test/MangaRipper.Test.csproj
@@ -54,29 +54,29 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MangaRipper.Batoto\MangaRipper.Batoto.csproj">
-      <Project>{0da63c57-4a0f-497e-9322-935af18506b1}</Project>
-      <Name>MangaRipper.Batoto</Name>
-    </ProjectReference>
     <ProjectReference Include="..\MangaRipper.Core\MangaRipper.Core.csproj">
       <Project>{13effe36-dbb8-4573-9260-b601177f66e3}</Project>
       <Name>MangaRipper.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MangaRipper.MangaFox\MangaRipper.MangaFox.csproj">
+    <ProjectReference Include="..\MangaRipper.Plugin.Batoto\MangaRipper.Plugin.Batoto.csproj">
+      <Project>{0da63c57-4a0f-497e-9322-935af18506b1}</Project>
+      <Name>MangaRipper.Plugin.Batoto</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\MangaRipper.Plugin.MangaFox\MangaRipper.Plugin.MangaFox.csproj">
       <Project>{f5d529b4-cbb6-46e2-9d32-25cef8365318}</Project>
-      <Name>MangaRipper.MangaFox</Name>
+      <Name>MangaRipper.Plugin.MangaFox</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MangaRipper.MangaHere\MangaRipper.MangaHere.csproj">
+    <ProjectReference Include="..\MangaRipper.Plugin.MangaHere\MangaRipper.Plugin.MangaHere.csproj">
       <Project>{b7f65da6-a435-46f8-8867-778a48fab8cf}</Project>
-      <Name>MangaRipper.MangaHere</Name>
+      <Name>MangaRipper.Plugin.MangaHere</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MangaRipper.MangaReader\MangaRipper.MangaReader.csproj">
+    <ProjectReference Include="..\MangaRipper.Plugin.MangaReader\MangaRipper.Plugin.MangaReader.csproj">
       <Project>{fe632cbc-ebbb-4509-b760-8ca031f87160}</Project>
-      <Name>MangaRipper.MangaReader</Name>
+      <Name>MangaRipper.Plugin.MangaReader</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MangaRipper.MangaShare\MangaRipper.MangaShare.csproj">
+    <ProjectReference Include="..\MangaRipper.Plugin.MangaShare\MangaRipper.Plugin.MangaShare.csproj">
       <Project>{a156fb0b-1e4a-4db0-b944-d23413ca8d58}</Project>
-      <Name>MangaRipper.MangaShare</Name>
+      <Name>MangaRipper.Plugin.MangaShare</Name>
     </ProjectReference>
   </ItemGroup>
   <Choose>
